fix: 60sec delay is triggered after 3 api requests while fetching financial statements
While fetching the financial statements, the condition of 60sec delay will only be checked after 3 api requests for each company. If the delay is started only after 5 api requests, this would be too late as the program will continue fetching data for the second company but 6 api requests are needed in total before the delay starts. So while fetching data for the second company, the program will be interrupted after sending the 6th api requests in a minute. Temporary solution would be to trigger the delay after every 3 api requests instead of 5.

Signed-off-by: Ooi Siang Tan <ooisiangtan@gmail.com>