feat: Print status code and 'Information' from response when API request call limit is reached
Previously as long as the response's status code and no 'symbol' key is found in the response,
user cannot differentiate what kind of error it was through the log message because only 'Something is wrong with the response' is printed.
Now, response status code and 'Information' from response will be printed if available.

Signed-off-by: Ooi Siang Tan <ooisiangtan@gmail.com>