fix: Handle empty database tables/pandas DataFrame
If a table is empty and user does not wish to update it while updating other tables, errors will be thrown due to empty dataframe.
Applied this fix to not perform operations on empty dataframes and save dataframes to db only when the table is selected in data_option
and it is not empty.

Signed-off-by: Ooi Siang Tan <ooisiangtan@gmail.com>